requirements.txt
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth>=2.25.0
grpcio>=1.59.0
google-api-core>=2.10.0
python-telegram-bot>=20.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
pdf2image>=1.17.0
python-dotenv>=1.0.0
requests>=2.31.0
